tests: bsim_bt: Add option to re-build incrementally
In the provided compile.sh script (which builds all bsim_bt
testcases used in the CI regression runs) add the option to
do an incremental build (if possible).
This is just a nicety for users who want to use this script
while testing locally.

This is controller with the existance of the variable INCR_BUILD

For ex., the script can be called, as:
 WORK_DIR=${ZEPHYR_BASE}/my_work_folder INCR_BUILD=y \
 tests/bluetooth/bsim_bt/compile.sh

+++ tests/bluetooth/bsim_bt/compile.sh
@@ -32,11 +32,16 @@ function compile(){
 
   echo "Building $EXE_NAME"
 
-  [ -d "${WORK_DIR}/${APP}" ] && rm ${WORK_DIR}/${APP} -rf
-  mkdir -p ${WORK_DIR}/${APP} && cd ${WORK_DIR}/${APP}
-  cmake -GNinja -DBOARD_ROOT=${BOARD_ROOT} -DBOARD=${BOARD} \
-        -DCONF_FILE=${CONF_FILE} ${CMAKE_ARGS} ${APP_ROOT}/${APP} \
-      &> cmake.out || { cat cmake.out && return 0; }
+  #Set INCR_BUILD when calling to only do an incremental build
+  if [ ! -v INCR_BUILD ] || [ ! -d "${WORK_DIR}/${APP}" ]; then
+      [ -d "${WORK_DIR}/${APP}" ] && rm ${WORK_DIR}/${APP} -rf
+      mkdir -p ${WORK_DIR}/${APP} && cd ${WORK_DIR}/${APP}
+      cmake -GNinja -DBOARD_ROOT=${BOARD_ROOT} -DBOARD=${BOARD} \
+            -DCONF_FILE=${CONF_FILE} ${CMAKE_ARGS} ${APP_ROOT}/${APP} \
+	  &> cmake.out || { cat cmake.out && return 0; }
+  else
+      cd ${WORK_DIR}/${APP}
+  fi
   ninja ${NINJA_ARGS} &> ninja.out || { cat ninja.out && return 0; }
   cp ${WORK_DIR}/${APP}/zephyr/zephyr.exe ${EXE_NAME}
